Cleaned up MainActivity.kt
