+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,10 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="button_red">#F45B69</color>
+    <color name="button_green">#92EF80</color>
+    <color name="button_yellow">#FFBE0B</color>
+    <color name="button_teal">#028090</color>
+    <color name="button_blue">#9CFFFA</color>
+    <color name="button_dark_green">#137547</color>
 </resources>
